+++ examples/nextjs-app/config.ts
@@ -5,8 +5,6 @@ import { mainnet, polygon, optimism, arbitrum } from 'wagmi/chains';
 export const config = createConfig(
   getDefaultConfig({
     appName: 'ConnectKit Next.js demo',
-    //infuraId: process.env.NEXT_PUBLIC_INFURA_ID,
-    //alchemyId:  process.env.NEXT_PUBLIC_ALCHEMY_ID,
     chains: [mainnet, polygon, optimism, arbitrum],
     walletConnectProjectId: process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID!,
   })