+++ packages/connectkit/src/components/ConnectKit.tsx
@@ -29,6 +29,7 @@ import { isFamily } from '../utils/wallets';
 import { useConnector } from '../hooks/useConnectors';
 import { WagmiContext, useAccount } from 'wagmi';
 import { Web3ContextProvider } from './contexts/web3';
+import { useChainIsSupported } from '../hooks/useChainIsSupported';
 
 export const routes = {
   ONBOARDING: 'onboarding',
@@ -201,16 +202,15 @@ export const ConnectKitProvider = ({
   useEffect(() => setErrorMessage(null), [route, open]);
 
   // Check if chain is supported, elsewise redirect to switches page
-  const { chain } = useAccount();
+  const { chain, isConnected } = useAccount();
+  const isChainSupported = useChainIsSupported(chain?.id);
+
   useEffect(() => {
-    if (
-      opts.enforceSupportedChains &&
-      Boolean(chain && !chains.some((x) => x.id !== chain?.id))
-    ) {
+    if (isConnected && opts.enforceSupportedChains && !isChainSupported) {
       setOpen(true);
       setRoute(routes.SWITCHNETWORKS);
     }
-  }, [chain, route, open]);
+  }, [isConnected, isChainSupported, chain, route, open]);
 
   // Autoconnect to Family wallet if available
   useEffect(() => {