+++ packages/connectkit/src/index.ts
@@ -1,4 +1,4 @@
-export const CONNECTKIT_VERSION = '1.7.0';
+export const CONNECTKIT_VERSION = '1.7.1';
 
 export * as Types from './types';
 export { default as getDefaultConfig } from './defaultConfig';
@@ -21,6 +21,7 @@ export { default as ChainIcon } from './components/Common/Chain';
 // Hooks
 export { default as useIsMounted } from './hooks/useIsMounted'; // Useful for apps that use SSR
 export { useChains } from './hooks/useChains';
+export { useChainIsSupported } from './hooks/useChainIsSupported';
 
 // TODO: Make this private
 export { default as ConnectKitModalDemo } from './components/ConnectModal/demo';